Fix user corrections in prompt
